Art-Net backend
This adds a backend to support Art-Net (https://art-net.org.uk/) which
is used by a number of larger LED lighting systems (I think the gateways
we have support 8192 channels or something).

There are a number of limitations, which I've tried to document in the
comments, so I'd currently call it experimental and I haven't put it in
the README yet. I'll hopefully have some time to improve this some more
over the next few weeks.